+++ src/main/java/me/jinjjahalgae/domain/proof/api/ProofController.java
@@ -1,7 +1,7 @@
 package me.jinjjahalgae.domain.proof.api;
 
 import lombok.RequiredArgsConstructor;
-import me.jinjjahalgae.domain.proof.usecase.interfaces.ProofCreateUseCase;
+import me.jinjjahalgae.domain.proof.usecase.interfaces.CreateProofUseCase;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class ProofController {
 
-    private final ProofCreateUseCase proofCreateUseCase;
+    private final CreateProofUseCase proofCreateUseCase;
