+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/interfaces/CreateProofUseCase.java
@@ -2,7 +2,7 @@ package me.jinjjahalgae.domain.proof.usecase.interfaces;
 
 import me.jinjjahalgae.domain.proof.dto.request.ProofCreateRequest;
 
-public interface ProofCreateUseCase {
+public interface CreateProofUseCase {
 
     void execute(ProofCreateRequest request, Long contractId);
 }
