+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/interfaces/CreateReProofUseCase.java
@@ -2,7 +2,7 @@ package me.jinjjahalgae.domain.proof.usecase.interfaces;
 
 import me.jinjjahalgae.domain.proof.dto.request.ProofCreateRequest;
 
-public interface ReProofCreateUseCase {
+public interface CreateReProofUseCase {
 
     void execute(ProofCreateRequest request, Long proofId);
 }
